requirements.txt
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
fonttools==4.58.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
